Modified : Ability to Find Maximum for 3 Floats FUnctionality Modified

+++ main/java/com/capgemini/assignment/findmaximum/FindMaximum.java
@@ -8,14 +8,14 @@ public class FindMaximum {
 	public static void main(String[] args) {
 		System.out.println("Welcome to Test Maximum");
 		 System.out.println("Enter three variables");
-	        int num1 = sc.nextInt();
-	        int num2 = sc.nextInt();
-	        int num3 = sc.nextInt();
+	        float num1 = sc.nextFloat();
+	        float num2 = sc.nextFloat();
+	        float num3 = sc.nextFloat();
 	        System.out.println("Maximum number is "+findMaximum(num1 , num2 , num3));
 	}
 
-	static Integer findMaximum(Integer x, Integer y, Integer z) {
-		int max=0;
+	static Float findMaximum(Float x, Float y, Float z) {
+		float max = 0;
 		if(x.compareTo(y)>0 && x.compareTo(z)>0) {
 			max=x;
 		}else if(z.compareTo(y)>0 && z.compareTo(x)>0) {
